Added small changes to existing codes and added new source files that will represent training phases
Made tweaks to existing code to improve its usage. Still need to add in
data to the base ship class related to being in or out of bounds, point
total, etc.

Added codes that will start diving into the training of the AI.
Starting out with a training implementation that will teach the AI a
controller to handle maneuvering the spaceship to desired states.